feat: chatMessage 공통 로직 useChatMessage에 분리
